update: 入口函数迁移重构
